test(termz): reach 100% coverage and dedup IsTerminal
Add tests for IsTerminal (previously untested) and for the *os.File
fall-through in IsColorTerminal, bringing the package to 100% statement
coverage on non-Windows.

IsTerminal was byte-identical in termz.go and termz_windows.go even
though term.IsTerminal is cross-platform. Move it to an unconstrained
file so there is a single implementation that the new test guards on
every platform, and split IsColorTerminal into termz_unix.go. The
Windows build keeps its console-mode IsColorTerminal.
